fix: Rodeo delivery (#101)
## Description
<!-- What issue or task does this change relate to? -->
this is an update to the transforms required to prepare the datasets in
the context of the RODEO project. New transforms keep the 3 variables in
the grib2 data rather than just tp


## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->
NA
## What issue or task does this change relate to?
<!-- link to Issue Number -->
NA

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->
NA

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

---------
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: b8raoult <53792887+b8raoult@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>