feat: General purpose clipping filter (#96)
## Description

This code defines a custom data filter plugin called Clipper for the
Anemoi framework. The purpose of the Clipper filter is to clip (limit)
the values of a specified field to a given range.

## Type of Change

-   [ ] Bug fix (non-breaking change which fixes an issue)
-   [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
-   [ ] Documentation update

## Issue Number

#84 

<!-- Alternatively, explain the motivation behind the changes and the
context in which they are being made. -->

## Code Compatibility

-   [x] I have performed a self-review of my code

### Code Performance and Testing

- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I ran the [complete Pytest
test](https://anemoi.readthedocs.io/projects/training/en/latest/dev/testing.html)
suite locally, and they pass

<!-- In case this affects the model sharding or other specific
components please describe these here. -->

### Dependencies

- [x] I have ensured that the code is still pip-installable after the
changes and runs
- [ ] I have tested that new dependencies themselves are
pip-installable.

<!-- List any new dependencies that are required for this change and the
justification to add them. -->

### Documentation

-   [x] My code follows the style guidelines of this project
- [ ] I have updated the documentation and docstrings to reflect the
changes
- [x] I have added comments to my code, particularly in
hard-to-understand areas

<!-- Describe any major updates to the documentation -->

## Additional Notes

<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

This is just a first draft that works both for 2d and 3d variables. It
would be nice to apply the same clipper to multiple variables at once,
if the required bounding values are the same. As it is, a new clipping
instance must be added for each variable to the dataset creation recipe.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Florian Pinault <floriankrb@users.noreply.github.com>
Co-authored-by: Florian Pinault <Florian.Pinault@ecmwf.int>