feat: Add SingleFieldFilter (#104)
## Description
Adds a `SingleFieldFilter` base class for creating filters which operate
on a single field at a time (those that are not required to be matching
filters).

## What problem does this change solve?
New feature - reduces the amount of new code required to add simple
filters and should help to improve consistency.

## What issue or task does this change relate to?
Supports #92 

##  Additional notes ##
Draft PR for feedback. Currently this branch is pointing at
`feat/testing-filters` for ease of comparison (builds off that branch,
assuming it will be merged in the future) – will eventually point at
`main`.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***


<!-- readthedocs-preview anemoi-transform start -->
----
📚 Documentation preview 📚:
https://anemoi-transform--104.org.readthedocs.build/en/104/

<!-- readthedocs-preview anemoi-transform end -->

---------

Co-authored-by: Baudouin Raoult <baudouin.raoult@ecmwf.int>
Co-authored-by: anaprietonem <ana.prietonemesio@ecmwf.int>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Michiel Van Ginderachter <michiel.vanginderachter@meteo.be>