feat: migrate sum filter (#121)
## Description
Moves the remaining `sum` filter from `anemoi-datasets` to
`anemoi-transform`. This is a direct lift & shift with the minimum
number of code changes to port the existing functionality across.

## What problem does this change solve?
Migrates filters over to anemoi-transform.

## What issue or task does this change relate to?
#92 

##  Additional notes ##
As noted in the tests, the filter does not currently support fields
across multiple levels, and it also flattens the output arrays – which
is different from many of the other filters. Updating/rationalising the
functionality of the filter will be carried out in a later PR - but this
should hopefully prevent breaking changes for users.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: anaprietonem <ana.prietonemesio@ecmwf.int>