fix: Warn rather than fail on metadata mismatch in matching filter (#126)
## Description
The
[change](https://github.com/ecmwf/anemoi-transform/pull/93/commits/01b05e0e9eb30a44cdbfcb0ff0264c13f20d4892)
introduced in #93 to raise an error on missing fields when using a
matching filter currently breaks the anemoi-inference tests. This PR
changes this to be a warning (and introduces a test to check it is
raised).

From @b8raoult: we use the same set of filters for inputs and forcings.
The solution would be to have different pre-processing per input.

## What problem does this change solve?
Bugfix

## What issue or task does this change relate to?
Introduced in #93

##  Additional notes ##
[See failed inference pytest
log](https://github.com/ecmwf/anemoi-inference/actions/runs/16219702370/job/45927280738)

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)