feat(variable): Period property for non-instant fields (#132)
## Description
When dealing with fields that are valid over a period (like averages) we
need to know what that period is. This is stored in the `period` key of
the variable metadata, but it was not directly exposed by the class.

This adds a period attribute to the Variable class that returns the
variable's valid period as a timedelta.

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)