test: change skip behaviour for rescale and add clear_step tests (#125)
## Description
Adds a test for the `clear_step` filter. Updates the skip behaviour when
`udunits2` is missing for the `rescale` filter to give clearer output
when the package is missing.

## What problem does this change solve?
Test update

## What issue or task does this change relate to?
n/a

##  Additional notes ##
n/a

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>