test: Add test for apply_mask filter (#136)
## Description
Adds a test for the `apply_mask` filter

## What problem does this change solve?
Improves test coverage

## What issue or task does this change relate to?
n/a

##  Additional notes ##
I'm not entirely sure of the value of essentially implementing a
parallel application of the actual masking operation, but I can see
value in adding a test despite this due to the rest of the surrounding
functionality (particularly being able to mock the loading of ancillary
data (i.e. masks in this case) which could be useful for other filters).

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)