feat: Add tox config (#131)
## Description
Adds a tox config to allow us to have CI-like behaviour locally (also
opening the door to potentially simplifying the GitHub Actions).

## What problem does this change solve?
Allows us to check that the pre-commit hooks pass, docs build
successfully, and that the tests pass in fresh environments across a
matrix of python versions (3.10 to 3.12 at the moment) locally. Also
included is the ability to test against release versions of the anemoi
dependencies, as well as using the versions at the head of main of the
respective packages (only utils for this repo).

## What issue or task does this change relate to?
n/a

##  Additional notes ##
n/a

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)

---------

Co-authored-by: Ana Prieto Nemesio <91897203+anaprietonem@users.noreply.github.com>