feat: Rodeo filters add docs and fix masking of QI (#145)
## Description
<!-- What issue or task does this change relate to? -->
Update the filters associated with the RODEO project to include docs and
tests. Additionally fix the masking to also be applied to QI.

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-transform start -->
----
📚 Documentation preview 📚:
https://anemoi-transform--145.org.readthedocs.build/en/145/

<!-- readthedocs-preview anemoi-transform end -->