docs: revisit documentation of filters (#169)
## Description

Now that all the filters have been migrated to this package, this is an
update of the way they are documented.

## What problem does this change solve?
<!-- Describe if it's a bugfix, new feature, doc update, or breaking
change -->

## What issue or task does this change relate to?
<!-- link to Issue Number -->

##  Additional notes ##
<!-- Include any additional information, caveats, or considerations that
the reviewer should be aware of. -->

***As a contributor to the Anemoi framework, please ensure that your
changes include unit tests, updates to any affected dependencies and
documentation, and have been tested in a parallel setting (i.e., with
multiple GPUs). As a reviewer, you are also responsible for verifying
these aspects and requesting changes if they are not adequately
addressed. For guidelines about those please refer to
https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the
[Contributor License
Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-transform start -->
----
📚 Documentation preview 📚:
https://anemoi-transform--169.org.readthedocs.build/en/169/

<!-- readthedocs-preview anemoi-transform end -->

---------

Co-authored-by: Aaron Hopkinson <197336788+aaron-hopkinson@users.noreply.github.com>